Make the sync transaction and error handler callbacks safe to run after the `SyncSession` has been destroyed.
The documentation for `sync::Session` states that they may be run after
the `sync::Session` goes away, which means they may also be run after the
`SyncSession` goes away. Capture a `weak_ptr` to `this` in the two
callback lambdas so we can detect whether the `SyncSession` exists and
guarantee it remains alive for the duration of the callback.